feat: add function to retrieve list of files from current folder
The recent code changes added a new function get_files_list() to the xtract-emails.py file. This function retrieves a list of files from the current folder that have a .json extension. This functionality enhances the script's capability to process data extraction.

Note: The suggested commit message adheres to the conventional format of starting with a type (e.g., feat, fix, chore) followed by a concise description of the changes made.